% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggplot_lm}
\alias{ggplot_lm}
\title{Helper function for easily plotting a linear regression model fit using polished ggplot and/or plotly aesthetics}
\usage{
ggplot_lm(
  lm_output,
  xvar_name = names(lm_output$model)[2],
  yvar_name = names(lm_output$model)[1],
  fill = NULL,
  alpha = NULL,
  shape = NULL,
  size = NULL,
  to_plotly = F
)
}
\description{
Helper function for easily plotting a linear regression model fit using polished ggplot and/or plotly aesthetics
}
\examples{
fit <- lm(Sepal.Length ~ Petal.Width, data=iris)
ggplot_lm(lm_output=fit)
ggplot_lm(lm_output=fit) + ggplot2::labs(title="Iris Regression", x="Petal Width (cm)", y="Sepal Length (cm)")
ggplot_lm(lm_output=fit, to_plotly=T)
}
